Fixes #1610: Move images from components to src
