+++ src/components/SkillCardList/SkillCardList.js
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom';
 import PropTypes from 'prop-types';
 import CircleImage from '../CircleImage/CircleImage';
 import styles from './SkillCardStyle';
-import StaffPick from '../images/staff_pick.png';
+import StaffPick from '../../images/staff_pick.png';
 import { urls } from '../../utils';
 
 function testExample(e, exampleText) {
